Detect already existing positions during sparse matrix node creation | Accept only non-zero values
If the user inputs a new value for the same position, it is logically wrong to create a new node  in the sparse matrix. It would be a different story if we were performing an operation, but here we are simply creating the matrix and so we should check if the given position already exists in the sparse matrix.

Also,the purpose of the sparse matrix would be defeated if we allow zeroes to be entered.